Fix bug with convertVideos function in actions/index.js
The ConvertPanel component was not passing in the list of videos to the convertVideos function, breaking the integration with the main process to receive the videoes from React.  Based on some feedback from the Udemy course questions, I added some bindings for the videos object array which is now being passed to the convertVideos function.